Introducing to the new EventService as module
