+++ Open-API/Open-MasterAPI/src/main/java/de/tammo/cloud/masterapi/CloudMasterAPI.java
@@ -6,12 +6,7 @@
 
 package de.tammo.cloud.masterapi;
 
-import de.tammo.cloud.masterapi.event.EventHandler;
-import lombok.Getter;
-
 public class CloudMasterAPI {
 
-	@Getter
-	private final EventHandler eventHandler = new EventHandler();
-
 }
+
