--- Open-API/Open-MasterAPI/src/main/java/de/tammo/cloud/masterapi/event/EventTarget.java
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2018, Open-Cloud-Services and contributors
- *
- * The code is licensed under the MIT License, which can be found in the root directory of the repository.
- */
-
-package de.tammo.cloud.masterapi.event;
-
-import java.lang.annotation.*;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface EventTarget {}
