+++ Open-Modules/Open-Event/src/main/java/de/tammo/cloud/event/Event.java
@@ -1,0 +1,3 @@
+package de.tammo.cloud.event;
+
+public class Event {}
