+++ Open-Modules/Open-Event/src/main/java/de/tammo/cloud/event/EventTarget.java
@@ -1,0 +1,7 @@
+package de.tammo.cloud.event;
+
+import java.lang.annotation.*;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface EventTarget {}
